fatture

+++ src/main/java/BuildWeekU5/EPIC/ENERGY/SERVICE/Repository/RuoloStatoFatturaidRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RuoloStatoFatturaidRepository extends JpaRepository<RuoloStatoFattura, Long> {
+public interface RuoloStatoFatturaidRepository extends JpaRepository<RuoloStatoFattura, Integer> {
 
 
 }
